Monkeypatch bar rate for reliable tests
